# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: XAI Blockchain API
  version: 1.1.0
  description: 'Production REST API for the XAI blockchain node.


    All endpoints require authenticated transport in production. Use the summaries
    and schemas in this document to integrate wallets, explorers, trading engines,
    and operational tooling.'
  contact:
    name: XAI Support
    url: https://xai-blockchain.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: http://localhost:5000
  description: Local development
- url: https://testnet-api.xai-blockchain.io
  description: Testnet
- url: https://api.xai-blockchain.io
  description: Production
...
paths:
...
  /mining/register:
    post:
      tags:
      - Mining Bonus
      summary: Register miner
      description: Registers a miner for XP/bonus tracking.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MiningRegisterInput'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  result:
                    type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /mining/achievements/{address}:
    get:
      tags:
      - Mining Bonus
      summary: Get achievements
      description: Returns unlocked achievements and progress for an address.
      parameters:
      - $ref: '#/components/parameters/AddressPathParam'
      - name: blocks_mined
        in: query
        schema:
          type: integer
          minimum: 0
      - name: streak_days
        in: query
        schema:
          type: integer
          minimum: 0
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  achievements:
                    type: array
                    items:
                      type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /mining/claim-bonus:
    post:
      tags:
      - Mining Bonus
      summary: Claim social bonus
      description: Claims a configured mining bonus such as social campaigns.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MiningBonusClaimInput'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  result:
                    type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /mining/referral/create:
    post:
      tags:
      - Mining Bonus
      summary: Create referral code
      description: Generates a referral code for a miner.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReferralCreateInput'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  result:
                    type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /mining/referral/use:
    post:
      tags:
      - Mining Bonus
      summary: Redeem referral code
      description: Registers a new miner using an existing referral code.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReferralUseInput'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  result:
                    type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /mining/user-bonuses/{address}:
    get:
      tags:
      - Mining Bonus
      summary: List user bonuses
      description: Returns all XP, streak, and referral bonuses tied to an address.
      parameters:
      - $ref: '#/components/parameters/AddressPathParam'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  bonuses:
                    type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /mining/leaderboard:
    get:
      tags:
      - Mining Bonus
      summary: Bonus leaderboard
      description: Returns the top addresses ranked by configured bonus metrics.
      parameters:
      - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  leaderboard:
                    type: array
                    items:
                      type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /mining/leaderboard/unified:
    get:
      tags:
      - Mining Bonus
      summary: Unified leaderboard
      description: Aggregated leaderboard blending XP, streaks, and referrals.
      parameters:
      - $ref: '#/components/parameters/LimitParam'
      - name: metric
        in: query
        schema:
          type: string
          default: composite
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  leaderboard:
                    type: array
                    items:
                      type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /mining/stats:
    get:
      tags:
      - Mining Bonus
      summary: Mining bonus stats
      description: System-wide statistics for the bonus engine.
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  stats:
                    type: object
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
...
components:
  parameters:
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
      description: Maximum number of records to return.
...
    AddressPathParam:
      name: address
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
        code:
          type: string
...
    MiningRegisterInput:
      type: object
      required:
      - address
      properties:
        address:
          type: string
    MiningBonusClaimInput:
      type: object
      required:
      - address
      - bonus_type
      properties:
        address:
          type: string
        bonus_type:
          type: string
    ReferralCreateInput:
      type: object
      required:
      - address
      properties:
        address:
          type: string
    ReferralUseInput:
      type: object
      required:
      - new_address
      - referral_code
      properties:
        new_address:
          type: string
        referral_code:
          type: string
        metadata:
          type: object
...
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: X-API-Key
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- api_key: []
- bearer_auth: []